nang cap wallet giu tien cho bid va autobid
- Chuyen wallet/top-up sang kieu long de xu ly so tien VND nhat quan.

- Them reserved_balance validation va wallet_transactions de log top-up/reserve/release/capture/refund.

- Tich hop AuctionService voi WalletService de giu tien nguoi dang dan dau, tha tien khi bi outbid, tru tien khi PAID va hoan tien khi CANCELED.

- Cap nhat client/server wallet top-up va cleanup du lieu wallet khi xoa user.

+++ client/src/main/java/userauth/gui/fxml/dialog/TopUpDialogController.java
@@ -98,9 +98,9 @@ public class TopUpDialogController {
             return;
         }
 
-        double amount;
+        long amount;
         try {
-            amount = UiInput.parsePositiveDecimal(amountText, "Top-up amount");
+            amount = (long) UiInput.parsePositiveDecimal(amountText, "Top-up amount");
         } catch (NumberFormatException ex) {
             showError(UiText.text("Invalid amount format."));
             return;
